+++ packages/dumi-plugin-mdx/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @secretflow/dumi-plugin-mdx
 
+## 0.3.0-alpha.1.0
+
+### Minor Changes
+
+- add transitional openapi features
+
+### Patch Changes
+
+- Updated dependencies
+  - @secretflow/unified-toolkit@0.4.0-alpha.1.0
+
 ## 0.2.0
 
 ### Minor Changes