+++ packages/unified-toolkit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @secretflow/unified-toolkit
 
+## 0.4.0-alpha.1.0
+
+### Minor Changes
+
+- add transitional openapi features
+
 ## 0.3.0
 
 ### Minor Changes